Match and link creverb.c

+++ src/musyx/runtime/CheapReverb/creverb.c
@@ -21,7 +21,7 @@ static void DLsetdelay(_SND_REVSTD_DELAYLINE* dl, s32 lag) {
 }
 
 static void DLcreate(_SND_REVSTD_DELAYLINE* dl, s32 len) {
-  dl->length = len * sizeof(f32);
+  dl->length = (s32)len * sizeof(f32);
   dl->inputs = (f32*)salMalloc(len * sizeof(f32));
   memset(dl->inputs, 0, len * sizeof(len));
   dl->lastOutput = 0.f;
@@ -62,7 +62,7 @@ bool ReverbSTDCreate(_SND_REVSTD_WORK* rv, float coloration, float time, float m
   rv->level = mix;
   rv->damping = damping;
   if (rv->damping < 0.05f) {
-    rv->damping = 0.05;
+    rv->damping = 0.05f;
   }
 
   rv->damping = 1.f - (rv->damping * 0.8f + 0.05f);
